core_app/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.2
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.4.4
marshmallow==3.19.0
marshmallow-sqlalchemy==0.28.0
requests==2.31.0
python-dotenv==1.0.0
bcrypt==4.0.1
psycopg2-binary==2.9.10
flask_cors==5.0.0
marshmallow-sqlalchemy==4.12.2